Handle Exception 28/10/2024

+++ Job-Recruitment/src/main/java/vn/trinhlam/jobhunter/controller/UserController.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -14,6 +15,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import vn.trinhlam.jobhunter.domain.User;
 import vn.trinhlam.jobhunter.service.UserService;
+import vn.trinhlam.jobhunter.service.error.IdInvalidException;
+
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.PutMapping;
 
@@ -36,8 +39,10 @@ public class UserController {
     }
 
     @DeleteMapping("/users/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable("id") long id) {
-
+    public ResponseEntity<String> deleteUser(@PathVariable("id") long id) throws IdInvalidException {
+        if (id >= 1000) {
+            throw new IdInvalidException("Id khong hop le");
+        }
         this.userService.delete(id);
         return ResponseEntity.status(HttpStatus.OK).body("deleted");
     }
